requirements.txt
pandas
numpy
googlemaps
requests
torch
torchvision
Pillow
transformers
geopy
psutil